Calculator App Readme
Description:

This is a simple Calculator app for Android devices that allows users to perform basic arithmetic operations such as addition, subtraction, multiplication, and division. The app provides a user-friendly interface with input fields and buttons for numbers and operations.

Features:

Addition: Add two numbers by entering them in the input fields and clicking the "+" button.
Subtraction: Subtract one number from another by entering them in the input fields and clicking the "-" button.
Multiplication: Multiply two numbers by entering them in the input fields and clicking the "x" button.
Division: Divide one number by another by entering them in the input fields and clicking the "/" button.
Refresh: Clear the input fields and reset the calculator by clicking the "Refresh" button.
How to Use:

Launch the Calculator app on your Android device.
You will see two input fields to enter numbers (editTextNumber1 and editTextNumber2) and an input field to enter the desired operation (editTextOperand).
Tap on the desired number buttons (0 to 9) to enter the numbers in the input fields. The input field with focus will be updated.
Tap on the operation buttons (+, -, x, /) to select the arithmetic operation.
If you need to perform multiple calculations, you can enter the next set of numbers and operation without refreshing the app.
To perform a calculation, enter two numbers in the input fields and select the desired operation. Tap the "=" button to view the result in the Answer TextView.
If you wish to clear the input fields and start a new calculation, click on the "Refresh" button.
Important Note:

Ensure that you have allowed installation from unknown sources in your Android device settings before installing the Calculator app.
This app is intended for basic arithmetic calculations and may not handle complex expressions or scientific calculations.
Technologies Used:

Java
Android Studio